refactor helper_functions
