+++ DokiApp/APIs/chat_apis.py
@@ -7,8 +7,15 @@ from rest_framework.permissions import IsAuthenticated
 from django.shortcuts import get_object_or_404
 from ..models import User, Chat, Message
 
-from ..Helper_functions.helper_functions import create_chat_name, oldest_unseen_message_id
 from ..Helper_functions.adapters import adapt_message, adapt_chat
+
+
+def create_chat_name(user_1_id, user_2_id):
+    name = str(user_1_id) + "_" + str(user_2_id)
+    if user_1_id > user_2_id:
+        name = str(user_2_id) + "_" + str(user_1_id)
+
+    return 'chat_' + name
 
 
 class ChatList(APIView):
@@ -35,13 +42,20 @@ class LoadOldChat(APIView):
         chat_name = create_chat_name(partner.id, request.user.id)
         chat = get_object_or_404(Chat, name=chat_name)
         messages = Message.objects.filter(chat=chat).order_by('-date')
-        oldest_unseen = oldest_unseen_message_id(messages, request.user)
+        oldest_unseen = self.oldest_unseen_message_id(messages, request.user)
 
         result = adapt_message(messages)
         self.set_messages_as_seen(result, request.user)
 
         return Response({"success": True, "messages": result, "oldest_unseen_message_id": oldest_unseen},
                         status=status.HTTP_200_OK)
+
+    def oldest_unseen_message_id(self, messages, user):
+        for message in messages.order_by('date'):
+            if (message.is_sender_doctor != user.is_doctor) and not message.seen:
+                return message.id
+
+        return 0
 
     def set_messages_as_seen(self, result, user):
         for message in Message.objects.filter(id__in=self.message_ids(result)):
