+++ DokiApp/APIs/reserve_apis.py
@@ -15,7 +15,7 @@ from rest_framework.response import Response
 from rest_framework.permissions import IsAuthenticated
 
 from ..Helper_functions.email_functions import send_reserve_message
-from ..Helper_functions.helper_functions import create_chat_name
+from .chat_apis import create_chat_name
 from ..permissions import IsPatient, IsDoctor
 from ..serializers import *
